fear: Cart Commands

+++ ShaliShop/src/Modules/CheckoutModule/src/CheckoutModule.Domain/Carts/DomainEvents/ItemQuantityUpdated.cs
@@ -1,3 +1,3 @@
 namespace CheckoutModule.Domain.Carts.DomainEvents;
 
-public record ItemQuantityUpdated(Guid CartId, Guid ProductId, int NewQuantity) : DomainEvent;
+public record ItemQuantityUpdated(Guid CartId, Guid ProductId, decimal NewQuantity) : DomainEvent;
