+++ ShaliShop/src/Modules/InventoryModule/src/InventoryModule.Domain/Inventories/DomainEvents/InventoryRestocked.cs
@@ -2,4 +2,4 @@ using Shared.Domain;
 
 namespace InventoryModule.Domain.Inventories.DomainEvents;
 
-public record InventoryRestocked(Guid InventoryId, Guid ProductId, int QuantityAdded) : DomainEvent;
+public record InventoryRestocked(Guid InventoryId, Guid ProductId, decimal QuantityAdded) : DomainEvent;
