+++ ShaliShop/src/Modules/InventoryModule/src/InventoryModule.Domain/Inventories/Repository/IInventoryRepository.cs
@@ -5,9 +5,9 @@ namespace InventoryModule.Domain.Inventories.Repository;
 
 public interface IInventoryRepository
 {
-    Task<Result> TryReserveStockAsync(Guid productId, int quantity, CancellationToken ct);
-    Task<Result> TryReleaseStockAsync(Guid productId, int quantity, CancellationToken ct);
-    Task<Result> RestockAsync(Guid productId, int quantity, CancellationToken ct);
+    Task<Result> TryReserveStockAsync(Guid productId, decimal quantity, CancellationToken ct);
+    Task<Result> TryReleaseStockAsync(Guid productId, decimal quantity, CancellationToken ct);
+    Task<Result> RestockAsync(Guid productId, decimal quantity, CancellationToken ct);
 
     Task<Inventory?> LoadAsync(Guid id, CancellationToken ct);
     Task SaveAsync(Inventory item, CancellationToken ct);
